Feat: 내보내기 dialg로의 전환 구현
