+++ app/src/main/java/com/example/egobook_frontent/ui/diary/view/DiaryExportDialogFragment.kt
@@ -1,0 +1,38 @@
+package com.example.egobook_frontent.ui.diary.view
+
+import android.content.DialogInterface
+import android.graphics.Color
+import android.graphics.drawable.ColorDrawable
+import android.os.Bundle
+import android.view.LayoutInflater
+import android.view.View
+import android.view.ViewGroup
+import androidx.fragment.app.DialogFragment
+import com.example.egobook_frontent.databinding.FragmentDiaryExportDialogBinding
+import com.example.egobook_frontent.removeScreenBlur
+
+class DiaryExportDialogFragment : DialogFragment() {
+
+    private var _binding: FragmentDiaryExportDialogBinding? = null
+    private val binding get() = _binding!!
+
+    override fun onCreateView(
+        inflater: LayoutInflater, container: ViewGroup?,
+        savedInstanceState: Bundle?
+    ): View {
+        dialog?.window?.setBackgroundDrawable(ColorDrawable(Color.TRANSPARENT))
+        _binding = FragmentDiaryExportDialogBinding.inflate(inflater, container, false)
+        return binding.root
+    }
+
+    override fun onCancel(dialog: DialogInterface) {
+        super.onCancel(dialog)
+        removeScreenBlur()
+        dismiss()
+    }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+}
