--- app/src/main/java/com/example/egobook_frontent/ui/diary/view/DiaryExportFragment.kt
@@ -1,27 +1,0 @@
-package com.example.egobook_frontent.ui.diary.view
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.DialogFragment
-import com.example.egobook_frontent.databinding.FragmentDiaryExportBinding
-
-class DiaryExportFragment : DialogFragment() {
-
-    private var _binding: FragmentDiaryExportBinding? = null
-    private val binding get() = _binding!!
-
-    override fun onCreateView(
-        inflater: LayoutInflater, container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentDiaryExportBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
